fix(falcon-front-kit): Renaming CheckoutLogic file to CheckoutContext
