+++ packages/falcon-front-kit/src/Checkout/index.ts
@@ -1,1 +1,1 @@
-export * from './CheckoutLogic';
+export * from './CheckoutContext';
